recommendations_service/requirements.txt
grpcio
psycopg2
pandas
scikit-learn
grpcio-tools
google
sentence_transformers